Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-angle-computed-expected.txt


PASS Property transform value 'rotate(min(1deg))' computes to 'matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)' 
FAIL Property transform value 'rotate(min(1grad))' computes to 'matrix(0.9998766324816606, 0.015707317311820675, -0.015707317311820675, 0.9998766324816606, 0, 0)' assert_equals: expected "matrix(0.9998766324816606, 0.015707317311820675, -0.015707317311820675, 0.9998766324816606, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
FAIL Property transform value 'rotate(min(1rad))' computes to 'matrix(0.5403023058681398, 0.8414709848078965, -0.8414709848078965, 0.5403023058681398, 0, 0)' assert_equals: expected "matrix(0.5403023058681398, 0.8414709848078965, -0.8414709848078965, 0.5403023058681398, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
FAIL Property transform value 'rotate(min(1turn))' computes to 'matrix(1, -2.4492935982947064e-16, 2.4492935982947064e-16, 1, 0, 0)' assert_equals: expected "matrix(1, -2.4492935982947064e-16, 2.4492935982947064e-16, 1, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
PASS Property transform value 'rotate(max(1deg))' computes to 'matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)' 
FAIL Property transform value 'rotate(max(1grad))' computes to 'matrix(0.9998766324816606, 0.015707317311820675, -0.015707317311820675, 0.9998766324816606, 0, 0)' assert_equals: expected "matrix(0.9998766324816606, 0.015707317311820675, -0.015707317311820675, 0.9998766324816606, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
FAIL Property transform value 'rotate(max(1rad))' computes to 'matrix(0.5403023058681398, 0.8414709848078965, -0.8414709848078965, 0.5403023058681398, 0, 0)' assert_equals: expected "matrix(0.5403023058681398, 0.8414709848078965, -0.8414709848078965, 0.5403023058681398, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
FAIL Property transform value 'rotate(max(1turn))' computes to 'matrix(1, -2.4492935982947064e-16, 2.4492935982947064e-16, 1, 0, 0)' assert_equals: expected "matrix(1, -2.4492935982947064e-16, 2.4492935982947064e-16, 1, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
PASS Property transform value 'rotate(min(1deg, 2deg))' computes to 'matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)' 
FAIL Property transform value 'rotate(min(1grad, 2grad))' computes to 'matrix(0.9998766324816606, 0.015707317311820675, -0.015707317311820675, 0.9998766324816606, 0, 0)' assert_equals: expected "matrix(0.9998766324816606, 0.015707317311820675, -0.015707317311820675, 0.9998766324816606, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
FAIL Property transform value 'rotate(min(1rad, 2rad))' computes to 'matrix(0.5403023058681398, 0.8414709848078965, -0.8414709848078965, 0.5403023058681398, 0, 0)' assert_equals: expected "matrix(0.5403023058681398, 0.8414709848078965, -0.8414709848078965, 0.5403023058681398, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
FAIL Property transform value 'rotate(min(1turn, 2turn))' computes to 'matrix(1, -2.4492935982947064e-16, 2.4492935982947064e-16, 1, 0, 0)' assert_equals: expected "matrix(1, -2.4492935982947064e-16, 2.4492935982947064e-16, 1, 0, 0)" but got "matrix(0.9998476951563913, 0.01745240643728351, -0.01745240643728351, 0.9998476951563913, 0, 0)"
PASS Property transform value 'rotate(max(1deg, 2deg))' computes to 'matrix(0.9993908270190958, 0.03489949670250097, -0.03489949670250097, 0.9993908270190958, 0, 0)' 
FAIL Property transform value 'rotate(max(1grad, 2grad))' computes to 'matrix(0.9995065603657316, 0.03141075907812829, -0.03141075907812829, 0.9995065603657316, 0, 0)' assert_equals: expected "matrix(0.9995065603657316, 0.03141075907812829, -0.03141075907812829, 0.9995065603657316, 0, 0)" but got "matrix(0.9993908270190958, 0.03489949670250097, -0.03489949670250097, 0.9993908270190958, 0, 0)"
FAIL Property transform value 'rotate(max(1rad, 2rad))' computes to 'matrix(-0.4161468365471424, 0.9092974268256817, -0.9092974268256817, -0.4161468365471424, 0, 0)' assert_equals: expected "matrix(-0.4161468365471424, 0.9092974268256817, -0.9092974268256817, -0.4161468365471424, 0, 0)" but got "matrix(0.9993908270190958, 0.03489949670250097, -0.03489949670250097, 0.9993908270190958, 0, 0)"
FAIL Property transform value 'rotate(max(1turn, 2turn))' computes to 'matrix(1, -4.898587196589413e-16, 4.898587196589413e-16, 1, 0, 0)' assert_equals: expected "matrix(1, -4.898587196589413e-16, 4.898587196589413e-16, 1, 0, 0)" but got "matrix(0.9993908270190958, 0.03489949670250097, -0.03489949670250097, 0.9993908270190958, 0, 0)"
FAIL Property transform value 'rotate(min(90deg, 0.26turn))' computes to 'matrix(6.123233995736766e-17, 1, -1, 6.123233995736766e-17, 0, 0)' assert_equals: expected "matrix(6.123233995736766e-17, 1, -1, 6.123233995736766e-17, 0, 0)" but got "matrix(0.9999897039488793, 0.00453784048117526, -0.00453784048117526, 0.9999897039488793, 0, 0)"
FAIL Property transform value 'rotate(min(1.57rad, 95deg))' computes to 'matrix(0.0007963267107332633, 0.9999996829318346, -0.9999996829318346, 0.0007963267107332633, 0, 0)' assert_equals: expected "matrix(0.0007963267107332633, 0.9999996829318346, -0.9999996829318346, 0.0007963267107332633, 0, 0)" but got "matrix(0.9996245977511435, 0.027398240287735538, -0.027398240287735538, 0.9996245977511435, 0, 0)"
PASS Property transform value 'rotate(max(91deg, 0.25turn))' computes to 'matrix(-0.017452406437283473, 0.9998476951563913, -0.9998476951563913, -0.017452406437283473, 0, 0)' 
FAIL Property transform value 'rotate(max(1.58rad, 90deg))' computes to 'matrix(-0.009203543268808559, 0.9999576464987401, -0.9999576464987401, -0.009203543268808559, 0, 0)' assert_equals: expected "matrix(-0.009203543268808559, 0.9999576464987401, -0.9999576464987401, -0.009203543268808559, 0, 0)" but got "matrix(6.123233995736766e-17, 1, -1, 6.123233995736766e-17, 0, 0)"
FAIL Property transform value 'rotate(min(270deg, max(0.25turn, 3.14rad)))' computes to 'matrix(-0.9999987317275395, 0.0015926529164868282, -0.0015926529164868282, -0.9999987317275395, 0, 0)' assert_equals: expected "matrix(-0.9999987317275395, 0.0015926529164868282, -0.0015926529164868282, -0.9999987317275395, 0, 0)" but got "matrix(0.998498672858271, 0.05477590985343363, -0.05477590985343363, 0.998498672858271, 0, 0)"
FAIL Property transform value 'rotate(max(0.25turn, min(270deg, 3.14rad)))' computes to 'matrix(-0.9999987317275395, 0.0015926529164868282, -0.0015926529164868282, -0.9999987317275395, 0, 0)' assert_equals: expected "matrix(-0.9999987317275395, 0.0015926529164868282, -0.0015926529164868282, -0.9999987317275395, 0, 0)" but got "matrix(0.998498672858271, 0.05477590985343363, -0.05477590985343363, 0.998498672858271, 0, 0)"
FAIL Property transform value 'rotate(calc(min(90deg, 1.58rad) + 0.25turn))' computes to 'matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)' assert_equals: expected "matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)" but got "matrix(-0.027572707274697524, 0.9996198006309919, -0.9996198006309919, -0.027572707274697524, 0, 0)"
FAIL Property transform value 'rotate(calc(min(90deg, 1.58rad) - 0.125turn))' computes to 'matrix(0.7071067811865476, 0.7071067811865475, -0.7071067811865475, 0.7071067811865476, 0, 0)' assert_equals: expected "matrix(0.7071067811865476, 0.7071067811865475, -0.7071067811865475, 0.7071067811865476, 0, 0)" but got "matrix(0.7263347879241293, -0.6873410913449087, 0.6873410913449087, 0.7263347879241293, 0, 0)"
FAIL Property transform value 'rotate(calc(min(90deg, 1.58rad) * 2)' computes to 'matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)' assert_equals: expected "matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)" but got "matrix(0.9984794916270877, 0.055124448297579676, -0.055124448297579676, 0.9984794916270877, 0, 0)"
FAIL Property transform value 'rotate(calc(min(90deg, 1.58rad) / 2)' computes to 'matrix(0.7071067811865476, 0.7071067811865475, -0.7071067811865475, 0.7071067811865476, 0, 0)' assert_equals: expected "matrix(0.7071067811865476, 0.7071067811865475, -0.7071067811865475, 0.7071067811865476, 0, 0)" but got "matrix(0.9999049456400823, 0.013787664214945474, -0.013787664214945474, 0.9999049456400823, 0, 0)"
PASS Property transform value 'rotate(calc(max(90deg, 1.56rad) + 0.25turn))' computes to 'matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)' 
PASS Property transform value 'rotate(calc(max(90deg, 1.56rad) - 0.125turn))' computes to 'matrix(0.7071067811865476, 0.7071067811865475, -0.7071067811865475, 0.7071067811865476, 0, 0)' 
PASS Property transform value 'rotate(calc(max(90deg, 1.56rad) * 2)' computes to 'matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)' 
PASS Property transform value 'rotate(calc(max(90deg, 1.56rad) / 2)' computes to 'matrix(0.7071067811865476, 0.7071067811865475, -0.7071067811865475, 0.7071067811865476, 0, 0)' 
FAIL Property transform value 'rotate(calc(min(90deg, 1.58rad) + max(0.25turn, 99grad)))' computes to 'matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)' assert_equals: expected "matrix(-1, 1.2246467991473532e-16, -1.2246467991473532e-16, -1, 0, 0)" but got "matrix(-0.18360823024919293, 0.9829995003990385, -0.9829995003990385, -0.18360823024919293, 0, 0)"
FAIL Property transform value 'rotate(calc(min(90deg, 1.58rad) - max(0.25turn, 99grad)))' computes to 'matrix(1, 0, 0, 1, 0, 0)' assert_equals: expected "matrix(1, 0, 0, 1, 0, 0)" but got "matrix(-0.12914174726147007, -0.9916261438234949, 0.9916261438234949, -0.12914174726147007, 0, 0)"
